# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bulk SMS Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DeviceStatus = typeof DeviceStatus[keyof typeof DeviceStatus];


export const DeviceStatus = {
  online: 'online',
  offline: 'offline',
  idle: 'idle',
} as const;

export interface Device {
  id: number;
  name: string;
  phoneNumber: string;
  status: DeviceStatus;
  /** @nullable */
  batteryLevel?: number | null;
  /** @nullable */
  signalStrength?: number | null;
  /** @nullable */
  lastSeen?: string | null;
  token: string;
  createdAt: string;
}

export interface DeviceInput {
  name: string;
  phoneNumber: string;
}

/**
 * @nullable
 */
export type DeviceHeartbeatStatus = typeof DeviceHeartbeatStatus[keyof typeof DeviceHeartbeatStatus] | null;


export const DeviceHeartbeatStatus = {
  online: 'online',
  offline: 'offline',
  idle: 'idle',
} as const;

export interface DeviceHeartbeat {
  /** @nullable */
  batteryLevel?: number | null;
  /** @nullable */
  signalStrength?: number | null;
  /** @nullable */
  status?: DeviceHeartbeatStatus;
}

export interface DeviceConnectInfo {
  token: string;
  connectUrl: string;
  qrData: string;
}

export interface Contact {
  id: number;
  name: string;
  phoneNumber: string;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  phoneNumber: string;
  /** @nullable */
  tags?: string | null;
}

export interface ContactUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  tags?: string | null;
}

export interface ContactList {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  contactCount: number;
  createdAt: string;
}

export interface ContactListDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  contactCount: number;
  contacts: Contact[];
  createdAt: string;
}

export interface ContactListInput {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface ListContactInput {
  contactId: number;
}

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  draft: 'draft',
  sending: 'sending',
  paused: 'paused',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Campaign {
  id: number;
  name: string;
  message: string;
  status: CampaignStatus;
  /** @nullable */
  deviceId?: number | null;
  /** @nullable */
  deviceName?: string | null;
  /** @nullable */
  contactListId?: number | null;
  /** @nullable */
  contactListName?: string | null;
  totalCount: number;
  sentCount: number;
  failedCount: number;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface CampaignInput {
  name: string;
  message: string;
  /** @nullable */
  deviceId?: number | null;
  /** @nullable */
  contactListId?: number | null;
  /** @nullable */
  scheduledAt?: string | null;
}

export type MessageStatus = typeof MessageStatus[keyof typeof MessageStatus];


export const MessageStatus = {
  queued: 'queued',
  sent: 'sent',
  failed: 'failed',
  delivered: 'delivered',
} as const;

export interface Message {
  id: number;
  /** @nullable */
  campaignId?: number | null;
  /** @nullable */
  campaignName?: string | null;
  /** @nullable */
  contactId?: number | null;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  deviceId?: number | null;
  /** @nullable */
  deviceName?: string | null;
  phoneNumber: string;
  /** @nullable */
  messageText?: string | null;
  status: MessageStatus;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
}

export interface DailyMessageCount {
  date: string;
  count: number;
}

export interface DashboardStats {
  devicesOnline: number;
  devicesTotal: number;
  messagesToday: number;
  successRateToday: number;
  activeCampaigns: number;
  totalContacts: number;
  messagesSentThisWeek?: DailyMessageCount[];
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  device_connected: 'device_connected',
  device_disconnected: 'device_disconnected',
  campaign_started: 'campaign_started',
  campaign_completed: 'campaign_completed',
  campaign_failed: 'campaign_failed',
  message_sent: 'message_sent',
  message_failed: 'message_failed',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  relatedId?: number | null;
  createdAt: string;
}

export type ListContactsParams = {
/**
 * @nullable
 */
listId?: number | null;
/**
 * @nullable
 */
search?: string | null;
};

export type ListMessagesParams = {
/**
 * @nullable
 */
campaignId?: number | null;
/**
 * @nullable
 */
status?: string | null;
};