# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bulk SMS Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: listId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Contact Lists ─────────────────────────────────────────────────────────────
  /contact-lists:
    get:
      operationId: listContactLists
      tags: [contact-lists]
      summary: List contact lists
      responses:
        "200":
          description: List of contact lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactList"
    post:
      operationId: createContactList
      tags: [contact-lists]
      summary: Create a contact list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactListInput"
      responses:
        "201":
          description: Created contact list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
  /contact-lists/{id}:
    get:
      operationId: getContactList
      tags: [contact-lists]
      summary: Get contact list with contacts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteContactList
      tags: [contact-lists]
      summary: Delete a contact list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contact-lists/{id}/contacts:
    post:
      operationId: addContactToList
      tags: [contact-lists]
      summary: Add a contact to a list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListContactInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contact-lists/{id}/contacts/{contactId}:
    delete:
      operationId: removeContactFromList
      tags: [contact-lists]
      summary: Remove contact from a list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Campaigns ────────────────────────────────────────────────────────────────
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign with stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/send:
    patch:
      operationId: sendCampaign
      tags: [campaigns]
      summary: Start sending a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/pause:
    patch:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a running campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/cancel:
    patch:
      operationId: cancelCampaign
      tags: [campaigns]
      summary: Cancel a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # ── Messages ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Contact:
      type: object
      required: [id, name, phoneNumber, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phoneNumber:
          type: string
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
    ContactInput:
      type: object
      required: [name, phoneNumber]
      properties:
        name:
          type: string
        phoneNumber:
          type: string
        tags:
          type: ["string", "null"]
    ContactUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]

    # ── ContactList ───────────────────────────────────────────────────────────
    ContactList:
      type: object
      required: [id, name, contactCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        contactCount:
          type: integer
        createdAt:
          type: string
    ContactListDetail:
      type: object
      required: [id, name, contactCount, contacts, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        contactCount:
          type: integer
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        createdAt:
          type: string
    ContactListInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
    ListContactInput:
      type: object
      required: [contactId]
      properties:
        contactId:
          type: integer

    # ── Campaign ──────────────────────────────────────────────────────────────
    Campaign:
      type: object
      required: [id, name, message, status, totalCount, sentCount, failedCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [draft, sending, paused, completed, cancelled]
        deviceId:
          type: ["integer", "null"]
        deviceName:
          type: ["string", "null"]
        contactListId:
          type: ["integer", "null"]
        contactListName:
          type: ["string", "null"]
        totalCount:
          type: integer
        sentCount:
          type: integer
        failedCount:
          type: integer
        scheduledAt:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    CampaignInput:
      type: object
      required: [name, message]
      properties:
        name:
          type: string
        message:
          type: string
        deviceId:
          type: ["integer", "null"]
        contactListId:
          type: ["integer", "null"]
        scheduledAt:
          type: ["string", "null"]

    # ── Message ───────────────────────────────────────────────────────────────
...