# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bulk SMS Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  name: string;
  phoneNumber: string;
  /** @nullable */
  tags?: string | null;
}